Split the converters into multiple modules
